claim_extraction_vision/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
openai
trafilatura
newspaper3k
lxml_html_clean
python-multipart
pillow
pytesseract
pytest
